Comment model (#4)
* user module completed

* zod validation implement for user create

* jwt implement and user update api created

* access token and refresh token and logout api created

* category create get and delete api created

* post model crud opetion completed

* comment model crud opetion completed

---------

+++ src/app/modules/category/category.service.ts
@@ -42,8 +42,12 @@ const updateCategory = async (id: string, payload: Partial<ICategory>) => {
   return updateCategory;
 };
 
-const deleteCategory = async (id:string) => {
-  await Category.findByIdAndDelete(id)
+const deleteCategory = async (id: string) => {
+  const existingCategory = await Category.findById(id);
+  if (!existingCategory) {
+    throw new Error("Category Not Found.");
+  }
+  await Category.findByIdAndDelete(id);
   return null;
 };
 
@@ -51,5 +55,5 @@ export const CategoryServices = {
   createCategory,
   getAllCategories,
   updateCategory,
-  deleteCategory
+  deleteCategory,
 };
